Feat: Implementando dialog de confirmação de exclusão de curso

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -6,6 +6,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/error-dialog.component';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -31,20 +32,27 @@ export class CoursesComponent implements OnInit{
   }
 
   onDelete(course: Course){
-    this.coursesService.delete(course._id).subscribe({
-      next: () => {
-        this.refresh()
-        this.snackBar.open("Curso deletado com sucesso!", 'X', { duration: 5000, verticalPosition: 'top', horizontalPosition: 'center'})
-      },
-      error: () => this.onError("Erro ao tentar remover curso!")
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: "Deseja realmente excluir o curso?",
     });
 
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if(result){
+        this.coursesService.delete(course._id).subscribe({
+          next: () => {
+            this.refresh()
+            this.snackBar.open("Curso deletado com sucesso!", 'X', { duration: 5000, verticalPosition: 'top', horizontalPosition: 'center'})
+          },
+          error: () => this.onError("Erro ao tentar remover curso!")
+        });
+      }
+    });
   }
 
   refresh(){
      this.courses$ = this.coursesService.list()
     .pipe(
-      catchError(error => {
+      catchError(() => {
         this.onError('Erro ao carregar cursos.')
         return of([])
       })
